requirements.txt
pystac_client==0.8.6
s2cloudless==1.7.3
h5py==3.14.0
loky==3.5.5
hydra-core==1.3.2
lightning~=2.5.2

boto3~=1.34
einops~=0.8.0
psycopg2-binary==2.9.9
geoalchemy2~=0.14.6
geopandas~=1.0.0
mgrs==1.5.0
numpy~=1.26.4
pandas~=2.2.2
pydantic~=2.9.2
pyproj~=3.7.1
pytest~=8.2.2
python-dotenv==1.0.1
rasterio~=1.4.3
requests~=2.32.3
rtree==1.2.0
scikit-learn==1.4.0
scipy~=1.14.0
shapely~=2.0.3
torch~=2.6.0
torchvision~=0.21.0
torchinfo~=1.8.0
tqdm~=4.66.4
wandb~=0.17.3

